backend/requirements.txt
# Recommend using Python 3.11
aiomysql
fastapi
bcrypt
passlib
pydantic
pyjwt
python-jose
python-multipart
uvicorn
python-dotenv